refactor(cli): address reviewer feedback for PR 3
Consolidate SearchModeDisplay, NavigationHelp, and NoResultsDisplay into a single SessionBrowserNav.tsx file. Rename NavigationHelp to NavigationHelpDisplay. Update SessionBrowser.tsx and SessionBrowserSearchNav.test.tsx accordingly. Delete redundant files.
